requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
diffusers==0.14.0
filelock==3.11.0
huggingface-hub==0.13.4
idna==3.4
importlib-metadata==6.3.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
packaging==23.0
Pillow==9.5.0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
sympy==1.11.1
tokenizers==0.13.3
torch==2.0.0
tqdm==4.65.0
transformers==4.27.4
typing_extensions==4.5.0
urllib3==1.26.15
zipp==3.15.0
